+++ src/types/segment.h
@@ -12,7 +12,7 @@ typedef struct {
     char gs_visited;
     int gs_distance_to_initial;
     int gs_distance_to_target;
-    int gs_factor;
+    unsigned long long int gs_factor;
     
     int frontiers_count;
     FrontierNode_t **frontiers;
